Updated README, Added SEMGREP CI
